add hover text on axis-chart bars
Move getAttrOrCP() from ember-devel.js to axis-1d.js, so it can be also used in axis-chart.js.
Split configureHorizTickHover() from axis-1d.js (based on an earlier version which is in axis-tracks.js, which was copied from draw-map.js) to form utils/hover.js

Add hoverTextFn(), configureChartHover(), based on axis-1d.js: hoverTextFn() and setupHover().
The axis-2d foreignObject currently overlays axis-chart masking the hover action; making it display:none is one way to access the bars hover;  later can use e.g. axis-accordion to separate these horizontally, for axis-chart the foreignObject is not used.
frontend/app/
 bf71726 16491 May 27 22:56  components/axis-chart.js
 8adf911 15121 May 27 22:40  components/draw/axis-1d.js
 46a1611  1328 May 27 22:59  utils/ember-devel.js
added :
 28f3788 1041 May 27 22:56 frontend/app/utils/hover.js
